Extract digest data fetch function out into own file.  🐿 v2.8.0
